Construindo Handle Para Classe do Tracker
